requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
flasgger==0.9.7
